Removed "raw" return type | Added pushNoti & checkNoti in case
As most of the data retrieved is decoded with respect to json, as because it makes the whole process a lot more easier. Thus the "raw" return type has been removed. Instead "json" is compulsory.

Added 2 new switch cases for the command for $_GET which is checkNoti(fication) and pushNoti(fication)