Fixed some object/array mix issue | Consistent returns | Added functions
There were parts where the variable was assigned simple values, but as the returned data is decoded in json, it has been changed to accommodate the value in json format.

Some array's identifier was not consistent, which is not anymore.

Added functions for pushNotification & checkNotification which pushes a notification in the client's device and the latter checks if the authorization was approved or denied.