Added functions | Changed HTTP response data handling | JSON dependency
Added further functionality to send notification to client's device and to check if it was approved or not.

HTTP response's data was handled on basis of an integer. Whereas what was returning was not integer at all. But json data. Thus it is now decoded and then values are assigned accordingly.

Added the JSON dependency to accommodate the JSON parse function.

Changed the format from .pwn -> .inc, as simply calling #include <tfasamp> wouldn't find the file.